Refactor PizzaCommandInvoker to be an FXMLComponent
