+++ src/main/java/org/jjspizzeria/jjspizzeria/command/Command.java
@@ -1,8 +1,9 @@
 package org.jjspizzeria.jjspizzeria.command;
 
 public abstract class Command {
-    abstract void execute();
-    abstract void undo();
+    public abstract void execute();
+    public abstract void undo();
+    
     // Topping commands will override this default behaviour to become undoable
     public boolean isUndoable() {
         return false;
